requirements-fixed.txt
# Vessels Platform - Fixed and Updated Dependencies
# Python 3.12+ required

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================

# Configuration & Validation
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0.1,<7.0.0

# Web Framework & API
flask>=3.1.0,<4.0.0
flask-cors>=4.0.0,<5.0.0
werkzeug>=3.0.0,<4.0.0

# Authentication & Security
PyJWT>=2.8.0,<3.0.0
cryptography>=44.0.0,<45.0.0
bcrypt>=4.1.0,<5.0.0

# Async Support
httpx>=0.26.0,<1.0.0
aiohttp>=3.9.0,<4.0.0

# Session Management
cachetools>=5.3.0,<6.0.0

# ============================================================================
# AI/ML STACK (Updated)
# ============================================================================

# Core ML
torch>=2.5.0,<3.0.0
transformers>=4.46.0,<5.0.0
sentence-transformers>=3.0.0,<4.0.0

# Embeddings & NLP
numpy>=1.26.0,<2.0.0
sentencepiece>=0.2.0,<1.0.0

# Speech (Optional - choose one)
faster-whisper>=0.10.0,<1.0.0  # Recommended

# LLM Inference (Optional)
llama-cpp-python>=0.2.27,<1.0.0

# ============================================================================
# DATABASE & STORAGE
# ============================================================================

# Graph Database
redis>=5.0.7,<6.0.0
# Note: FalkorDB installed separately as Redis module

# GraphRAG
graphiti-core>=0.3.0,<1.0.0

# ============================================================================
# UTILITIES
# ============================================================================

# HTTP Clients
requests>=2.32.0,<3.0.0

# Data Processing
python-dateutil>=2.8.2,<3.0.0

# Job Scheduling (for background tasks)
apscheduler>=3.10.0,<4.0.0

# Retry Logic
tenacity>=8.3.0,<9.0.0

# JSON Performance
orjson>=3.9.0,<4.0.0

# ============================================================================
# DEVELOPMENT & TESTING
# (Install with: pip install -r requirements-dev.txt)
# ============================================================================
# Moved to requirements-dev.txt

# ============================================================================
# REMOVED DEPENDENCIES (Unused - saved 120MB)
# ============================================================================
# beautifulsoup4==4.12.3  # NOT USED
# lxml==5.1.0  # NOT USED
# networkx==3.2.1  # NOT USED (have FalkorDB for graphs)
# pandas==2.1.4  # NOT USED
# scikit-learn==1.3.2  # NOT USED
# psycopg2-binary==2.9.9  # NOT USED (SQLite + Redis only)
# matplotlib==3.8.2  # NOT USED
# seaborn==0.13.1  # NOT USED

# ============================================================================
# OPTIONAL DEPENDENCIES
# ============================================================================
# Install with: pip install vessels[tts]
# [tts]
# kokoro-tts>=0.1.0
# piper-tts>=1.2.0

# Install with: pip install vessels[distributed]
# [distributed]
# petals>=2.0.0

# Install with: pip install vessels[nostr]
# [nostr]
# nostr-sdk>=0.1.0
